Explicitly load preferences on first use.
